Skeleton mock SAS HTTP server adapter and SAS client adapter.
This change includes a skeleton of a mock SAS HTTP server which is
manually configured to forward requests based on the SAS-CBSD and
SAS-SAS URL patterns to appropriate stub methods in the mock object.

This code will be extended to directly assist in developing the CBSD
tests by being extended to mock the SAS side of the protocol, and
in developing the SAS tests by providing a mock during test development.

The change also contains a skeleton of a SAS-CBSD and SAS-SAS client
class. This class is suitable for making requests of a SAS endpoint
serving the SAS-SAS and SAS-CBSD protocols. This code will be extended
to provide protocol compliance tests, be useful as a mock during
test development, and be able to be manipulated for automated testing.