Updates to CA test processes
This commit updates the openssl.cnf file with new sections for each of
the different types of certs in the WInnForum hierarchy. The
configuration uses the version of the ComSec document approved/published
by WG2.
This version of the configuration includes the WInnForum registered OIDs
for ROLE as identified/defined in the ComSec document.
There is also an addition of a PowerShell script. This script should be
able to be run on a Windows machine that has OpenSSL installed. It will
first prompt for your OpenSSL /bin directory path and the name of your
OpenSSL config filename. From there, it should create in your /bin
folder 11 private keys, CSRs, and signed certificates. There are
numerous variables in the .ps1 file itself which can be modified to
update what types of certs are created.
Last, this commit also adds a folder labeled "ExamplePKI" which contains
the certificate files which result from running the PowerShell script.