Add python interface to ITM model code.
This change adds a python callable module which interfaces to the C++
ITM code. It includes a setup.py file which works with distutils.
To build and install the module, use 'python setup.py build_ext -i'
to build the object file into the local directory.

The module can then be used following the test code in test_itm.py
That example shows how to import and use the point-to-point module
exported by the C++ code and tests a single run of the model
corresponding to one of the C++ test code vectors.