Update how "ROLE" OIDs are encoded
This commit primarily:
1) Changes the "ROLE" OIDs
Instead of storing the role value as an arbitrary extension with text
values for each role, this commit stores them in the Certificate
Policies Extension.
In order to do that effectively, I broke out the currently 6 possible
roles into a sub-OID structure of the previous ROLE OID
(1.3.6.1.4.1.46609.1.1.1-1.3.6.1.4.1.46609.1.1.6)
2) Adds the OpenSSL configuration and script sections necessary to
create the CBSD OEM intermediate and EE cert.

I also cleaned up a lot of the variables in the PowerShell script and
OpenSSL conf file.