Add a script to retrieve NED elevation map files.
Retrieves base tileset zip files and updates. According to the USGS
NED website, the longer filenames are updates to the base tileset.
These are frequently replacing existing tiles, but may also augment
the dataset by filling in gaps in the coverage.

The coverage is currently (fall 2016) over the continental US, HI,
and many US territories. There are some missing areas in AK.