Added effective heights calculation
The comparison to 200 m is based on effective heights instead of actual heights. Modified code accordingly. ***Note: profile[] is now calculated as many as three times in this code for each path. Should we just declare it as a global variable and calculate once at the beginning, or implement another way to do it so we're not creating the same profile three times?