Reference implementation for the Winnforum propagation models (ITM and Hybrid) (#277)
* Reference implementation for the Winnforum propagation models (ITM and Hybrid).

This relies on backend propagation ITS reference implementation (in
C++) and some python extension module wrappers. Minor modifications to
the ITS code have been done for supporting WinnForum extensions and for
performance.
Optimized Geo modules are provided for accessing geo data and computing
great circles.

See README files for details.

* Add missing files: .gitignore and NED prj file.

* Miscellaneous improvements, and bug fix in order of eHata profile.

 - Improve Python extension wrappers to capture wrong inputs as
exceptions (ehata_its_py.cpp and itm_its_py.cpp)
 - Reverse order of profile for case of eHata (bug fix)
 - Move the CbsdEffectiveHeight() routine and tests in ehata.py
 - Add makefiles
 - Improve documentation

* Add CBSD indoor mode support to both models.

Following request that indoor extra loss of 15dB is applied as part
of the 2 models instead of within the client code.

* Improvements to the documentation and build process.

- Added top level Makefile.
- Added instructions for compiling in Windows.
- Improved distutils setup scripts for Windows building.

* Add NLCD reference unit test data.

* Minor cleanup.

Function name change for nlcd, and better descriptin of building
C code in Windows.

* Add detection in NlcdVote of out of tile.

Will now raise an exception by default in that case of codes=0.

* Address Red comments 3 and 4.

 - C3: Fix Hybrid between 100m and 1km. Ehata part should use
 fixed distance of 1km
 - C4: removed the refractivity and climate from function
 prototypes, to reduce possibility of errors from module clients.

* Fix eHata median-to-mean formula.

* Modify slightly the effective height calculation.

The idea is that its is computed strictly in the interval
[3km-15km], and not taking in account first point below 3km.

* Change the Hybrid mode over 80km method.

Original implementation proposed to reuse a subset of the full
profile to avoid a new profile extraction exactly at 80km.
However doing this may create discontinuous issues, as the J
compensation factor (which is supposed to insure continuity from
below 80km to above 80km) may change drastically due to:
 - the profile sampling points will change from one  point to an other
 given that we only keep 1500pts in the profile and make the resolution
 vary.
 - so the J factor may be varying for any points above 80km, especially
 given the high non linearity of the ITM models. The test_over_80km
 for example demonstrate such behavior. The J calculated at 80km is 0.
 The J calculate using profile at 81km cut to almost 8km shows J=13dB

* Fix NED terrain driver issues.

 - indexing using float index crash in newer numpy version
 - scalar version was always reloading the tile for each access
 caused by an issue withwrong key cached.

* Modify documentation for new geo data location.

As part of the integration of the reference geo data
in the SAS-data directory, the new geo data (NLCD and NED)
location is in the top level data/geo/ned and /data/geo/nlcd
folders (this is to avoid corruption in multi-step PR, as well
as enabling the possibility to have the SAS-data repository as
as a submodule of this repository).

* Add global script for running all unit tests

* Fix average pathloss calculation to be linear.

The mean pathloss shall be computed in linear space,
not in dB space.