Apply same fix in the Python code for Effective height.
Distance exactly equal to 3km would produce a Nan otherwise.
Also make sure that the step is robust against passing integer value in
the profile.

Note that we change back to the int() instead of ceil() to be coherent
with the internal E-Hata calculation.
All this wont make a big change in practice. At worse we include one
sample below between 2970m and 3000m3 in the calculation of the average
terrain.