Disable useless internals return from prop model.
The returned internals is not used for anything in test harness - it
is more used for study of the model themselves.
It contains a big profile array, and will prevent efficient caching
of the propagation result.
As such I disable the return of these internals by default.
This is transparent to test harness.