FSS blocking: add the frequency check that FSS higher frequency always >= 3700MHz.
This test is present in IAP, but not in aggregate interference code.
For consistency, added it here as well.
Note that if this condition is not the case, then the blocking
protection methodology of looking at the left side would be "physically"
wrong, but this would be a matter for WG1 specification body.