Fix MCP by computing ESC interference post-filter (#697)
* Integrate ESC mask into aggregate interference.

The ESC mask for the 3650-3680MHz currently not taken into account in a
straightforward way.
There was some discussion one month ago (between Yi, Sam, and Navin mainly
in thread: 'ESC sensor protection discussion') about the processing
required for ESC in IAP and aggregate interference.

Currently the IAP introduce a special roll-off attenuation on those
channels, and the interference is computed like if no roll off exist.
The aggregate interference is reusing the interference.py routines, ie
ignore entirely the filter mask. Normally for consistency it should be
taken into account. It was mentioned in the email thread that by 'hacking'
the protection levels for those out of band channels, the test could be
made consistent. Those protection levels are done in MCP and it is very
hacky to do it there.

All this is similar to FSS filter mask used in FSS blocking.
To me it seems more straightforward to simply compute the interference
using the ESC mask, as expected.

In this PR, I introduce it in the context of aggregate_interference
only (and not in the IAP process). This is to feed the discussion.
Next stpe would be to:
 - do it all the time when computing interference for ESC
 - removing the special IAP code with the roll-off
It seems that it would simplify greatly the code.

NOTE: I did not update the aggregate_interference_test unit-test, and
will wait for approval on the process.

* Change code to do everything for ESC at filter output.

This makes things more consistent with other processing such
as FSS blocking.
It also fixes failing MCP when no peer SAS as the check are done
against the abolute threshold at filter output, although the threshold
should adapted per frequency.