Remove the use of the mp Manager, using the pool mechanism instead.
Some profiling of the Manager as used in the code shows that it does not
scale at all, as it constantly fork new process and lock during
contention when updating the shared dict (plus IPC overheads).
As a consequence, beyond a few processes , the actual running time is
constant or actually increase. All gains from parallel calculation is
lost on the shared objects management, which can take up to 80% of the
time for 10 process for example.

Using simply the regular returnes value from pool.map is much more
scalable, in addition of being simple.