Delete redundant config sections, and add CRL DPs.
When configs are duplicated between [ foo ] and [ foo_sign ], delete
the former.  This information is written to the *.csr file, ignored
(because we don't set copy_extensions), and finally deleted by
"rm *.csr".  Reducing redundancy makes the code easier to read/edit.

Add a CRL Distribution Point to all non-root certificates.
The following are particularly important for SxS_16:
- revoked_cbsd_ca  -> [ cbsd_req_sign ]
- revoked_sas_ca   -> [ sas_req_sign ]
- revoked_proxy_ca -> [ oper_req_sign ]