Refactor assertTlsHandshake{Succeed,Failure} into a common function.
Prior to this commit, assertTlsHandshakeFailure was missing a call to
load_verify_locations(). The CRL tests (SxS_11 and SxS_16) simply
detected an unknown server certificate, and didn't progress to the point
of detecting a revoked client certificate.

By checking for success and failure using a common function, we can be
more confident that the test is correct.

Fixes #488