Renew grants with every heartbeat in functional tests
We agreed long ago in test dev that we would set grantRenew=True in each heartbeat request made during the functional test cases, thus allowing the SAS Admin to set a relatively short grant duration (important for reducing the runtime of HBT.6) without unintended consequences (e.g. the grant expiring in the middle of MCP.1). This PR adds this functionality, which was inadvertently omitted.