Fix Fine Rolling Hill correction.
When the rolling hill is below 10m, the correction should be zero.
Otherwise it does not follow the EHata spec on one hand, plus it
introduces some randomness and discontinuity in the correction of
2dB or 4dB when the terrain is very flat near the mobile (ie a
microscopic change of altitude can cause up to 4dB variation in
pathloss).

This change is make to reflect the proposed change in the ITS original
code base, which was integrated recently at:
https://github.com/NTIA/ehata/commit/b54b71f5d2d46579afbc3b8cc01f952002943746