EHata: Cleanup Average terrain code for the NTIA compatibility mode. (#821)
* Cleanup Average terrain code for NTIA compatibility mode.

The legacy mode replicating ITS implementation has been upgraded
to latest version found in NTIA reference version (as was done in
https://github.com/NTIA/ehata/commit/80954b1914944841584f5ca34d30afcc24596ffb).

This has no effect on test harness which used WinnforumExtensions=True,
but removes confusion when comparing the Winnforum version versus the
NTIA version.

* Upgrade unit test for change.

The unit test are performed using the compatibility mode
(Winnforum_Extension=False). Since a bug is fixed in that mode,
update accordingly the unit tests.