Add support for 'Best-of-N' reference move list.
By default OFF. The computeMoveList can now compute several
move list and take the best (best being the smallest one, ie the
one with biggest keep list, and so producing the largest interference
hopefully).

Note that this proposal slow down the reference model by factor N.
The idea is to avoid "outliers" by using a small N, even something as
small as 2 to 5.
All previous DPA margin studies are demonstrating the issue by comparing
the SAS UUT creating a "best" move list (not a crazy assumption for an
efficient SAS), to the "worst case" move list reference model across many runs.
This can happen in practice in a random fashion. But if we repeat the
move list calculation a few time and take the smallest one, the chance
for it to happen is much lower. For example if we want to avoid the 10%
worst move list:
 - nominally we have 1 chance over 10 to fall in this scenario
 - by choosing the best of 3, the probability fall to 1 over 1000.

Note: this is only loosely related to Michael Souryal "move list
intersection" proposal which involves more move list calculation but
ends up with a "composite move list" capturing the true geometry of
the problem (but that composite move list may not be a valid move list
in the SAS protection sense).