Add "best of N" feature to applicable IPR tests.
Within the DPA configuration (the same place where 'movelistMargin' is specified), there is now an optional field called 'bestOfN'. If not defined, the test harness will use the default number (one). If defined, it should be an integer value >= 1. See example in IPR_7 default config (all other configs are unchanged).