Review the way to obtain proper keep-list.
Previous method is not reliable when no peer sas, and
if FAD dump changes slight;y the coordinates by rounding.
So decided to introduce proper way which is to get the nbor list
from a new specialized routine in move_list.py.
Right now it has been decided to use that only for logging, so
we are sure we do not introduce any regression in test harness, given
the upcoming code freeze.