Properly support grants within DPA.
For such grants, we filter them out before running
regular move list algorithm, then we include them back
on the proper move list.

This method insures that we would get same move list results as
using an extra exclusion zone on the DPA: hence if this is done,
we will not see a result variation.

Note that another proposed method is to keep the inside grants
in the move list process, but simply making sure that we add them at
the end (if not added by the move list process).
Doing this will not however give very stable results as those inside
grants will affect the aggregated interference, in a sensitive manner
depending strongly on the actual distribution of protected points.