DPA: properly support grants within  DPA (#869)
Simply make sure that the "inside DPA" grants are included in the proper move list
after the regular processing:

for inband channel, include all the grant which overlaps
for OOB, include all the grants
This method insures that we would get same move list results as if we were adding
protection points exactly under each CBSD.

Note that initial implementation was proposing to first filter out the grants inside the DPA and then run
the move list on the remaining outside grants. This is somewhat a bit more optimal in theory,
although we do not expect a major difference in practice. It has the disadvantage of changing
a bit more the philosophy of the move list algorithm, and thus requires handling a special case.

Adding unit test on the logic implemented.
And for sake of improving DPA unit test, added some more unit test which was part of
the dpa_mgr_example.py